requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
llama-cpp-python>=0.3.0
huggingface_hub>=0.20.0
requests>=2.25.0
